+++ changelog.md
@@ -1,6 +1,6 @@
 Notice: All 1.7.x changes are present in 2.0.x aswell
 
-# v1.7.0-rc5 (next)
+# v1.7.0-rc5
 - [FEATURE] sync() now correctly returns with an error when foreign key constraints reference unknown tables
 - [BUG] sync() no longer fails with foreign key constraints references own table (toposort self-dependency error)
 - [FEATURE] makes it possible to specify exactly what timestamp attributes you want to utilize [#1334](https://github.com/sequelize/sequelize/pull/1334)